Split tria_description.h.
